feat: [SPLT-1549] Adiciona pacote v2 na trip-schema
