+++ trip-schema/src/main/java/global/maplink/trip/schema/v2/problem/TripRequest.java
@@ -1,4 +1,4 @@
-package global.maplink.trip.schema.problem;
+package global.maplink.trip.schema.v2.problem;
 
 import global.maplink.emission.schema.EmissionRequest;
 import global.maplink.freight.schema.FreightCalculationRequest;
@@ -6,7 +6,6 @@ import global.maplink.place.schema.PlaceRouteRequest;
 import lombok.*;
 
 import java.time.LocalDate;
-import java.util.Date;
 import java.util.List;
 import java.util.Set;
 
@@ -18,7 +17,7 @@ import static lombok.AccessLevel.PRIVATE;
 @NoArgsConstructor(force = true, access = PRIVATE)
 public class TripRequest {
     private final List<SitePoint> points;
-    private final String calculationMode;
+    private final CalculationMode calculationMode;
     private final Set<String> restrictionZones;
     private final Set<String> avoidanceTypes;
     private final TollRequest toll;
