+++ trip-schema/src/test/java/global/maplink/trip/schema/v2/problem/SitePointTest.java
@@ -1,6 +1,7 @@
-package global.maplink.trip.schema.problem;
+package global.maplink.trip.schema.v2.problem;
 
 import global.maplink.json.JsonMapper;
+import global.maplink.trip.schema.v2.problem.SitePoint;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
