+++ trip-schema/src/test/java/global/maplink/trip/schema/v2/problem/TollRequestTest.java
@@ -1,8 +1,9 @@
-package global.maplink.trip.schema.problem;
+package global.maplink.trip.schema.v2.problem;
 
 import gloabl.maplink.toll.schema.TollConditionBillingType;
 import gloabl.maplink.toll.schema.TollVehicleType;
 import global.maplink.json.JsonMapper;
+import global.maplink.trip.schema.v2.problem.TollRequest;
 import org.junit.jupiter.api.Test;
 
 import static global.maplink.trip.testUtils.ProblemSampleFiles.TOLL_REQUEST;
