+++ trip-schema/src/test/java/global/maplink/trip/schema/v2/solution/CrossedBorderResponseTest.java
@@ -1,6 +1,7 @@
-package global.maplink.trip.schema.solution;
+package global.maplink.trip.schema.v2.solution;
 
 import global.maplink.json.JsonMapper;
+import global.maplink.trip.schema.v2.solution.CrossedBorderResponse;
 import lombok.Data;
 import org.junit.jupiter.api.Test;
