+++ trip-schema/src/test/java/global/maplink/trip/schema/v2/solution/SolutionLegTest.java
@@ -1,4 +1,4 @@
-package global.maplink.trip.schema.solution;
+package global.maplink.trip.schema.v2.solution;
 
 import gloabl.maplink.toll.schema.*;
 import global.maplink.geocode.schema.Address;
@@ -7,7 +7,8 @@ import global.maplink.place.schema.Category;
 import global.maplink.place.schema.LegResult;
 import global.maplink.place.schema.PlaceRoute;
 import global.maplink.place.schema.SubCategory;
-import global.maplink.trip.schema.problem.SitePoint;
+import global.maplink.trip.schema.v2.problem.SitePoint;
+import global.maplink.trip.schema.v2.solution.SolutionLeg;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
